Parsing the file still throws an error on blanks
Seems like on some blank files, it is still throwing an error when trying to parse the INI file.  `$CurrrentIniFileContent = Parse-IniFile -IniFile $SourceIniFile`

I propose a change to add a try/catch block.  If it can't parse the ini file, then just delete it from there, then it will recreate it with desired settings.